refactor: Major refactor - class BinaryString
Per my session with Kim Barnes during the last week of January, I
completed a major refactor of this code to use a new class, BinaryString.
BinaryString has a value that is a string of ones and zeros (hence the name).
These strings were being passed around all over the place and there were lots
of operations being made on them so they deserved their own class. The class
now implements operations such as circular_left_shift and complement.

The refactor was done by starting at the lowest level classes (ChFunction,
MajFunction, and ParityFunction) and replacing the strings there with instances
of the new class. I then worked up through the classes until the Processor class
was finally complete.

A lot of the tests had to be modified because the methods were now expecting the
BinaryString class instead of just regular ruby strings.